doc-intelligence/pyproject.toml
[project]
name = "doc-intelligence"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-ai-documentintelligence>=1.0.2",
    "azure-core>=1.35.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
